remove "using boost" to avoid ambiguous use of make_shared
C++11 (supported by clang on MacOS 10.9) includes std::make_shared,
so a bare "make_shared" is ambiguous if we are both "using std" and
"using boost".  Could have just added a namespace to "make_shared",
but removed "using boost" to avoid this kind of thing happening
again with something else.